requirements.txt
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
fastapi
uvicorn
python-multipart
tensorflow-cpu>=2.4.0